requirements.txt
pandas
supabase
python-dotenv
autopep8
